# wasmparser: improve visit_operator performance