--- a/src/lib.rs
+++ b/src/lib.rs
@@ -297,14 +297,8 @@ impl OutputArg {
 }
 
 pub fn parse_binary_wasm(parser: wasmparser::Parser, bytes: &[u8]) -> Result<()> {
+    let mut allocs = wasmparser::OperatorsReaderAllocations::default();
     for payload in parser.parse_all(&bytes) {
-        parse_payload(payload)?;
-    }
-    return Ok(());
-
-    fn parse_payload(
-        payload: Result<wasmparser::Payload, wasmparser::BinaryReaderError>,
-    ) -> Result<()> {
         match payload? {
             wasmparser::Payload::TypeSection(s) => parse_section(s)?,
             wasmparser::Payload::ImportSection(s) => parse_section(s)?,
@@ -321,11 +315,12 @@ pub fn parse_binary_wasm(parser: wasmparser::Parser, bytes: &[u8]) -> Result<()>
                 for item in locals.by_ref() {
                     let _ = item?;
                 }
-                let mut ops = locals.into_operators_reader();
+                let mut ops = locals.into_operators_reader(std::mem::take(&mut allocs));
                 while !ops.eof() {
                     ops.read()?;
                 }
                 ops.finish()?;
+                allocs = ops.into_allocations();
             }
 
             wasmparser::Payload::InstanceSection(s) => parse_section(s)?,
@@ -343,8 +338,8 @@ pub fn parse_binary_wasm(parser: wasmparser::Parser, bytes: &[u8]) -> Result<()>
 
             _ => (),
         }
-        Ok(())
     }
+    return Ok(());
 
     fn parse_section<'a, T>(s: wasmparser::SectionLimited<'a, T>) -> Result<()>
     where