… crates/wasmparser/src/binary_reader.rs
… crates/wasmparser/src/readers/core/operators.rs