This PR addresses some of the performance issues discussed in #2180 and #2202.

- When validating, it eliminates the need for the reader to have its own stack of frames or a depth count. The BinaryReader has a `visit_operator` function again, where the visitor implements the new `FrameStack` trait so that `visit_operator` can get the current frame's kind. Validation uses this to avoid duplicating the FrameStack logic between the reader and validator.

- For consumers who don't want to implement the `FrameStack` trait, they can use an `OperatorsReader` which keeps an internal `FrameStack`, but it can now be created from its allocations (and turned back into them), avoiding the need to **allocate** a stack of frames each time an `OperatorsReader` is constructed.

For validating lots of tiny modules and one big modules, I'm seeing these absolute timings on an Intel Core i7-9700T @ 4.3 GHz, running this command: `cargo bench -p wasmparser --bench benchmark -- --baseline pre-change 'validate/(tests|spidermonkey)'`. Percentages are relative to the first row.

| commit | validate/tests | validate/spidermonkey |
| --- | --- | --- |
| 90c156ff15a81cb0360fb949e6504a4799d3b5cc (just before #2134) |  5.1770 ms| 39.707 ms|
| 0354dde10184d7ff813282a2837d9a4ed8e766e6 (PR #2134) |  5.2136 ms (+0.7%)| 39.929 ms (+0.6%) |
| 351f1537490f7d5e4a61f8ab16302984589d25d2 (v1.229.0, #2149) | 5.1867 ms (+0.2%) | 39.474 ms (-0.6%) |
| a05eaa578ab6de49782a70d6f42a79b6e8548d6e (current main branch) | 5.7693 ms (+11.4%) | 41.529 ms (+4.6%) |
| e3c880c612186fab71ad84c6601c66da79a27250 (this PR) | 5.7307 ms (+10.7%) | 40.691 ms (+2.4%) |

The measurements suggest there may be some other slowdowns between the v1.229 release and the current main branch; this may be the best I can do with regard to reading operators, but I think there are probably also other opportunities to close the gap.