@alexcrichton 2025-06-16
Ok I've now gotten a chance to sit down with this and review it more closely, thanks for the patience. Also cc @Robbepop in case you hadn't already seen this.

I'd agree though that tracking the `validate/tests` benchmark over time isn't too useful since that is the whole test suite which includes new tests written, new spec tests, etc, so the set of what's being tested can fluctuate. Perhaps it should just be removed entirely from the benchmark suite...

Otherwise though I'm curious to dig in a bit more to the spidermonkey numbers. My overall stance on this is that I'm not a fan of making things less ergonomic in the name of performance but if it needs to be done it needs to be done. In that sense if this is a clear and unambiguous win then I think something like it should land, but I'm effectively double-checking that. In that sense I'm wary of the absolute numbers here in that Criterion is a wall-time-based benchmarking system which are generally notoriously unreliable when it comes to small timings, and 40ms here is IMO a pretty small amount of time (making it difficult for Criterion to measure changes).

For example the numbers you collected show that your change in this PR is a 2.2% improvement over the current `main` banch. However the numbers also show that the 1.233.0 release was a 2.3% improvement over 1.232.0. However in this release [practically nothing changed](https://github.com/bytecodealliance/wasm-tools/compare/v1.232.0...v1.233.0) related to core wasm parsing/validation (some component bits but those are unrelated). In that sense this aligns with what I'd roughly expect in that in the millisecond range Criterion, on the same code, can have +/- 5% swings (ish).

Interestingly the biggest regression you've measured, 3.7% in 1.230 -> 1.231 [actually landed a few optimizations](https://github.com/bytecodealliance/wasm-tools/compare/v1.230.0...v1.231.0) such as https://github.com/bytecodealliance/wasm-tools/pull/2197. I'm not sure how best to explain that...

Basically I'm overall a bit skeptical of the numbers collected here, but at the same time I can also naively see how this would probably improve performance given that there's fewer data structures in play. I'd prefer to not have to land this in the sense that it regresses ergonomics and makes things a bit more difficult to use, but that's also not the end of the world and something that can be figured out regardless.  I'm otherwise basically lamenting how the benchmarking infrastructure for wasmparser is not great and I at least personally feel I still don't have a great handle on what the regressions are...