@Robbepop 2025-06-16
@alexcrichton thanks for the cc!

This might not really have anything to do with performance, but while trying to use the new `wasmparser` version in Wasmi, I noticed that the newly introduced [`TypedSelectMulti`](https://docs.rs/wasmparser/latest/wasmparser/enum.Operator.html#variant.TypedSelectMulti) operator is weird in two ways:

1. It is the only `Operator` that is not actually a valid Wasm operator.
2. It is the only `Operator` that has a non-`Copy` field. (This gave me some minor headache.)

Do we really want to have such a "weird" operator? To me this indicates that `wasmparser`'s parser and validator have an unclear border or grey area in between them since recent changes.

> I'm otherwise basically lamenting how the benchmarking infrastructure for wasmparser is not great and I at least personally feel I still don't have a great handle on what the regressions are...

This sounds as if we should not attempt to optimize anything until we have better data (or benchmarks).