@keithw 2025-06-17
@alexcrichton Thanks, I appreciate these thoughtful comments. I'm not particularly attached to this PR; if there is a concern about a performance hit from the state introduced in #2134, this eliminates that state completely, without affecting conformance to chapter 5 (which as you know I am attached to). If there turns out not to be a significant performance concern, I'm equally fine sticking with the status quo. Given these results, I do suspect the initial focus on #2134 may have been a false alarm (with regard to end-to-end validation time at least), but I'm only doing Intel benchmarks here. I have yet to see absolute numbers on the `validate/spidermonkey` benchmark on arm64 but let me ping @Robbepop on the other thread.

From looking at a perf report, one of the contributors to performance seems to be that if we move the `visit_operator` implementation back into `binary_reader.rs` (as this PR does), calls to things like `read_var_u32` are more likely to get inlined vs. when `visit_operator` is implemented in `crates/wasmparser/src/readers/core/operators.rs`. This is maybe a <2% effect but I think it is real. (These calls do get inlined if we mark `read_var_u32` `#[inline(always)]`.)

The rest might just be the play of chance with regard to any small change going through LLVM on a non-PGO build. But obviously you should run the numbers yourself and see what you think.

One idea that you might like would be adding some more "intensive" benchmark workloads that might be more statistically robust, e.g. a validation workload that alternates validating spidermonkey.wasm and bz2.wasm 100 times in a row. If you want this, happy to send a PR but you probably know what you want here.